requirements.txt
Django==2.2.2
mysqlclient==1.4.2.post1
djangorestframework==3.9.4
coreapi==2.3.3
django-filter==2.1.0
social-auth-app-django==3.1.0
python-dateutil==2.8.0
